requirements.txt
Flask
pandas
dnspython
PyPDF2
openpyxl
xlrd
